backend/tests/requirements.txt
# Testing Dependencies for Content Generation Platform
# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-html==4.1.1
pytest-benchmark==4.0.0

# HTTP and API testing
httpx==0.25.0
requests-mock==1.11.0
aioresponses==0.7.4

# Database testing
pytest-postgresql==5.0.0
pytest-redis==3.0.2
sqlalchemy-utils==0.41.1
alembic==1.12.1

# Test containers for integration testing
testcontainers[postgresql]==3.7.1
testcontainers[redis]==3.7.1

# Load and performance testing
locust==2.17.0
pytest-timeout==2.1.0
memory-profiler==0.60.0
psutil==5.9.6

# Mock and fixtures
factory-boy==3.3.0
faker==19.12.0
responses==0.23.3
freezegun==1.2.2

# Security testing
bandit==1.7.5
safety==2.3.5

# Code quality and analysis
pylint==3.0.2
flake8==6.1.0
black==23.10.1
isort==5.12.0
mypy==1.6.1

# Documentation testing
pydocstyle==6.3.0

# Workflow testing
n8n-python-client==0.1.0

# Monitoring and metrics testing  
prometheus-client==0.18.0

# AI/ML testing utilities
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.1

# Data validation
pydantic==2.4.2
marshmallow==3.20.1

# Logging and debugging
structlog==23.2.0

# Environment management
python-dotenv==1.0.0

# Utility libraries
click==8.1.7
rich==13.6.0
tabulate==0.9.0